requirements.txt
numpy~=1.21.5
torch~=1.11.0
matplotlib~=3.5.1
torchvision~=0.12.0
opencv-python~=4.5.5.64
sty~=1.0.4
tqdm~=4.62.3
pandas~=1.2.3
Pillow~=9.0.1
seaborn~=0.11.2
scipy~=1.7.3
scikit-image~=0.19.2
scikit-learn~=1.1.0